+++ lib/Model/Payment.php
@@ -58,6 +58,8 @@ use YandexCheckout\Model\PaymentMethod\AbstractPaymentMethod;
  * @property Metadata $metadata Метаданные платежа указанные мерчантом
  * @property CancellationDetailsInterface $cancellationDetails Комментарий к отмене платежа
  * @property CancellationDetailsInterface $cancellation_details Комментарий к отмене платежа
+ * @property AuthorizationDetailsInterface $authorizationDetails Данные об авторизации платежа
+ * @property AuthorizationDetailsInterface $authorization_details Данные об авторизации платежа
  */
 class Payment extends AbstractObject implements PaymentInterface
 {
@@ -143,6 +145,13 @@ class Payment extends AbstractObject implements PaymentInterface
      * @since 1.0.13
      */
     private $_cancellationDetails;
+
+    /**
+     * Данные об авторизации платежа
+     * @var AuthorizationDetailsInterface
+     * @since 1.0.18
+     */
+    private $_authorizationDetails;
 
 
     /**
@@ -521,7 +530,6 @@ class Payment extends AbstractObject implements PaymentInterface
         return $this->_cancellationDetails;
     }
 
-
     /**
      * Устанавливает комментарий к статусу canceled: кто отменил платеж и по какой причине
      * @param CancellationDetailsInterface $value Комментарий к статусу canceled
@@ -529,5 +537,23 @@ class Payment extends AbstractObject implements PaymentInterface
     public function setCancellationDetails(CancellationDetailsInterface $value)
     {
         $this->_cancellationDetails = $value;
+    }
+    /**
+     * Возвращает данные об авторизации платежа
+     * @return AuthorizationDetailsInterface|null Данные об авторизации платежа
+     * @since 1.0.18
+     */
+    public function getAuthorizationDetails()
+    {
+        return $this->_authorizationDetails;
+    }
+
+    /**
+     * Устанавливает данные об авторизации платежа
+     * @param AuthorizationDetailsInterface $value Данные об авторизации платежа
+     */
+    public function setAuthorizationDetails(AuthorizationDetailsInterface $value)
+    {
+        $this->_authorizationDetails = $value;
     }
 }