+++ tests/Model/PaymentData/PaymentDataGooglePayTest.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Tests\YandexCheckout\Model\PaymentData;
+
+use YandexCheckout\Model\PaymentData\PaymentDataGooglePay;
+use YandexCheckout\Model\PaymentMethodType;
+
+require_once __DIR__ . '/AbstractPaymentDataGooglePayTest.php';
+
+class PaymentDataGooglePayTest extends AbstractPaymentDataGooglePayTest
+{
+    /**
+     * @return PaymentDataGooglePay
+     */
+    protected function getTestInstance()
+    {
+        return new PaymentDataGooglePay();
+    }
+
+    /**
+     * @return string
+     */
+    protected function getExpectedType()
+    {
+        return PaymentMethodType::GOOGLE_PAY;
+    }
+}