+++ tests/Request/Payments/Payment/CreateCaptureRequestSerializerTest.php
@@ -29,12 +29,12 @@ class CreateCaptureRequestSerializerTest extends TestCase
             foreach ($options['receiptItems'] as $item) {
                 $expected['receipt']['items'][] = array(
                     'description' => $item['title'],
-                    'quantity' => empty($item['quantity']) ? 1 : $item['quantity'],
+                    'quantity' => $item['quantity'],
                     'amount' => array(
                         'value' => $item['price'],
                         'currency' => isset($options['currency']) ? $options['currency'] : CurrencyCode::RUB,
                     ),
-                    'vat_code' => empty($item['vatCode']) ? $options['taxSystemCode'] : $item['vatCode'],
+                    'vat_code' => $item['vatCode'],
                 );
             }
             if (!empty($options['receiptEmail'])) {
@@ -68,10 +68,6 @@ class CreateCaptureRequestSerializerTest extends TestCase
                             'quantity' => Random::int(1, 10),
                             'price' => Random::int(100, 100),
                             'vatCode' => Random::int(1, 6),
-                        ),
-                        array(
-                            'title' => Random::str(10),
-                            'price' => Random::int(100, 100),
                         ),
                     ),
                     'receiptEmail' => Random::str(10),